time: make real timestamps 128 bit

+++ arch/x86_64/vdso/time.c
@@ -13,7 +13,7 @@ static uint64_t boot_time_syscall(void) {
     return SYSCALL0(SYSCALL_GET_NANOSECONDS_SINCE_BOOT).integer;
 }
 
-static int64_t real_time_syscall(void) {
+static __int128_t real_time_syscall(void) {
     return real_time_from_boot_time(boot_time_syscall());
 }
 
@@ -21,7 +21,7 @@ static uint64_t boot_time_kvmclock(void) {
     return x86_64_read_kvmclock(&vdso_info.arch.kvmclock) - vdso_info.arch.time_offset;
 }
 
-static int64_t real_time_kvmclock(void) {
+static __int128_t real_time_kvmclock(void) {
     return real_time_from_boot_time(boot_time_kvmclock());
 }
 
@@ -29,7 +29,7 @@ static uint64_t boot_time_tsc(void) {
     return timeconv_apply(vdso_info.arch.tsc2ns_conv, x86_64_read_tsc()) - vdso_info.arch.time_offset;
 }
 
-static int64_t real_time_tsc(void) {
+static __int128_t real_time_tsc(void) {
     return real_time_from_boot_time(boot_time_tsc());
 }
 
@@ -44,7 +44,7 @@ static uint64_t (*resolve_boot_time(void))(void) {
 
 __attribute__((ifunc("resolve_boot_time"))) EXPORT uint64_t hydrogen_boot_time(void);
 
-static int64_t (*resolve_real_time(void))(void) {
+static __int128_t (*resolve_real_time(void))(void) {
     switch (vdso_info.arch.time_source) {
     case X86_64_TIME_SYSCALL: return real_time_syscall;
     case X86_64_TIME_KVMCLOCK: return real_time_kvmclock;
@@ -53,4 +53,4 @@ static int64_t (*resolve_real_time(void))(void) {
     }
 }
 
-__attribute__((ifunc("resolve_real_time"))) EXPORT int64_t hydrogen_get_real_time(void);
+__attribute__((ifunc("resolve_real_time"))) EXPORT __int128_t hydrogen_get_real_time(void);
